requirements.txt
gdown==3.12.2
torch==1.7.1
torch-cluster==1.5.9
torch-geometric==1.6.3
torch-scatter==2.0.7
torch-sparse==0.6.9
torch-spline-conv==1.2.1
torchvision==0.8.2
tensorboard==2.5.0
scipy==1.5.4
scikit-fmm==2019.1.30
scikit-image==0.15.0
scikit-learn==0.22.2.post1
networkx
msgpack-numpy==0.4.7.1
numpy
numpy-quaternion
imageio
matplotlib
opencv-python
